/login, /logout, /gifthistory, /claim2fa, /fortniteaccess, /winterfest

+++ src/main/java/com/tb24/discordbot/Session.kt
@@ -20,7 +20,6 @@ import com.tb24.fn.util.EAuthClient
 import com.tb24.fn.util.getPreviewImagePath
 import net.dv8tion.jda.api.EmbedBuilder
 import net.dv8tion.jda.api.Permission
-import net.dv8tion.jda.api.entities.Message
 import org.greenrobot.eventbus.Subscribe
 import org.greenrobot.eventbus.ThreadMode
 import org.slf4j.Logger
@@ -68,11 +67,11 @@ class Session @JvmOverloads constructor(val client: DiscordBot, val id: String, 
 	fun login(source: CommandSourceStack?, fields: Map<String, String>, auth: EAuthClient = EAuthClient.FORTNITE_ANDROID_GAME_CLIENT, sendMessages: Boolean = true): Int {
 		if (source != null) {
 			val grantType = fields["grant_type"]
-			if (grantType != "device_auth" && grantType != "device_code" && source.guild?.selfMember?.hasPermission(Permission.MESSAGE_MANAGE) == true && sendMessages) {
+			if (grantType != "device_auth" && grantType != "device_code" && source.hasMessage && source.guild?.selfMember?.hasPermission(Permission.MESSAGE_MANAGE) == true && sendMessages) {
 				source.message.delete().queue()
 			}
 			if (api.userToken != null) {
-				logout(if (sendMessages) source.message else null)
+				logout()
 			}
 			if (grantType != "device_code") {
 				if (sendMessages) {
@@ -94,13 +93,9 @@ class Session @JvmOverloads constructor(val client: DiscordBot, val id: String, 
 	}
 
 	@Synchronized
-	fun logout(message: Message? = null) {
-		val logoutMsg = message?.run {
-			if (author.idLong == author.jda.selfUser.idLong) {
-				editMessage(Utils.loadingText("Logging out")).complete()
-			} else {
-				channel.sendMessage(Utils.loadingText("Logging out")).complete()
-			}
+	fun logout(source: CommandSourceStack? = null) {
+		val logoutMsg = source?.run {
+			channel.sendMessage(Utils.loadingText("Logging out")).complete()
 		}
 		var logoutSuccess = false
 		try {
